refactor(database): move blockverify to internal/, dedupe migration check
Cubic's review found verifyBlockSelfConsistent in
internal/integration/benchmark_test.go duplicated blockverify.Hash's
logic, since database/plugin/blob/internal/blockverify's Go-internal
scope only reached packages under database/plugin/blob -- not
internal/integration, which needed the identical check for its
migration-fixture test. The two copies could silently drift.

Moved the package to internal/blockverify (top-level, reachable from
anywhere in the module) and updated database/plugin/blob/{aws,gcs}'s
imports accordingly. internal/integration's verifyBlockSelfConsistent
now calls blockverify.Hash directly instead of maintaining a second
copy.
