+++ database/plugin/blob/aws/content_verification_test.go
@@ -20,7 +20,7 @@ import (
 	"testing"
 
 	"github.com/aws/aws-sdk-go-v2/service/s3"
-	"github.com/blinklabs-io/dingo/database/plugin/blob/internal/blockverify"
+	"github.com/blinklabs-io/dingo/internal/blockverify"
 	gledger "github.com/blinklabs-io/gouroboros/ledger"
 	lcommon "github.com/blinklabs-io/gouroboros/ledger/common"
 	"github.com/blinklabs-io/ouroboros-mock/fixtures"
